main.mjs has been modularized to import other functions
