+++ src/index.html
@@ -50,7 +50,7 @@
 
 
 
-    <script src="scripts/main.mjs"></script>
+    <script type="module" src="scripts/main.mjs"></script>
 </body>
 
 </html>
